[artifactory] move full images query to lib function.
Add get-full-images command that queries multiple images for all their
tags and queries each tag for the metadata.json contents and file
metadata and returns the combined list. This is the same thing that the
artifactory-docker-query command was doing but moved into the
viasat.apis.artifactory namespace so that it can be used by other things.