[cfn-desc-all] script to describe all stacks (not just running ones).
The AWS DescribeStacks call only shows information for running stacks.
The ListStacks call will give all stacks including deleted stacks but
the information is very limited. In order to get full details for all
stacks, the script does ListStacks and then uses the ARN/IDs to call
DescribeStacks for one stack at a time which is not efficient but seems
to be the only way to really do that.