[aws] Set role session name to somethine more memorable than "session1"
Used to uniquely (well, sortof) a session when the same role is assumed
by different principals for different reasons. e.g the username is
appended when a role is assumed from the web console. At least this
gives a bit of a clue as to who assumed the session when
troubleshooting via the CloudTrail logs.