[artifactory] Throw if user/token are not provided
Not setting artifactory environment variables would trigger an
interactive prompt for the user and/or token, which causes problems in
CI (or when users are piping outputs somewhere else).

To fix this, the artifactory ns now more strictly throws if user/token
are not available, so that by default we get a loud error in contexts
like CI.

We move the human-friendly interactive prompt out to the
artifactory-docker-query script, where it's potentially still
nice/helpful. Similarly, we move dotenv loading of env vars out to the
artifactory-docker-query script.

Finally, we start to standardize the various secret names to just
ARTIFACTORY_IDENTITY_TOKEN, no longer supporting ARTIFACTORY_PASSWORD
or ARTIFACTORY_API_KEY.