[viasat.apis.github] add helpful warning messages.
If the endpoint or action arg doesn't refer to an actual endpoint or
action in the Octokit API then return an error and list the options.

Add a check that all parameters needed by the resolved URL are present
in the user's command line arguments. If not, fatal with an error
listing the missing args.