Add viasat.apis.docker and docker-query command.
This library/command queries the standard Docker Hub API. The command
takes two parameters: namespace and image. For Docker official packages,
use "library" as the namespace (this is default for general docker
commands but is required for this comand).

It does not currently support authentication (for private repos and
higher rate limits).

Also, improve the base-url defaulting for viasat.apis.npm.