Support artifactory bearer token. Drop prompting.
If no username value is set then assume that the token is a bearer token
and pass it that way.

Remove interactive prompting for artifactory credentials. It complicates
the logic for bearer vs user+token. Also, the dependencies are a bit
heavy it's probably overkill for this type of tool. The creds can still
be set in environment or via CLI parameters.